% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{apply_jacobian}
\alias{apply_jacobian}
\title{Evaluate Jacobian of model.}
\usage{
apply_jacobian(
  url,
  out_wrt,
  in_wrt,
  parameters,
  vec,
  config = jsonlite::fromJSON("{}")
)
}
\arguments{
\item{url}{URL the model is running at.}

\item{out_wrt}{Output variable to take Jacobian with respect to.}

\item{in_wrt}{Input variable to take Jacobian with respect to.}

\item{parameters}{Model input parameter (a list of vectors).}

\item{vec}{Vector to multiply Jacobian by.}

\item{config}{Model-specific configuration options.}
}
\value{
Jacobian with respect to given input and output variables, applied to given vector.
}
\description{
Evaluate Jacobian of model.
}
