% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{get_models}
\alias{get_models}
\title{Get models supported by server.}
\usage{
get_models(url)
}
\arguments{
\item{url}{URL the model is running at.}
}
\value{
List of models supported by server.
}
\description{
Get models supported by server.
}
