% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{protocol_version_supported}
\alias{protocol_version_supported}
\title{Check if model's protocol version is supported by this client.}
\usage{
protocol_version_supported(url)
}
\arguments{
\item{url}{URL the model is running at.}
}
\value{
TRUE if model's protocol version is supported by this client, FALSE otherwise.
}
\description{
Check if model's protocol version is supported by this client.
}
