% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{supports_apply_hessian}
\alias{supports_apply_hessian}
\title{Check if model supports Hessian action.}
\usage{
supports_apply_hessian(url, name)
}
\arguments{
\item{url}{URL the model is running at.}

\item{name}{Name of the desired model.}
}
\value{
TRUE if model supports Hessian action, FALSE otherwise.
}
\description{
Check if model supports Hessian action.
}
