% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{supports_gradient}
\alias{supports_gradient}
\title{Check if model supports gradient evaluation.}
\usage{
supports_gradient(url)
}
\arguments{
\item{url}{URL the model is running at.}
}
\value{
TRUE if model supports gradient evaluation, FALSE otherwise.
}
\description{
Check if model supports gradient evaluation.
}
