% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbridge.R
\name{model_output_sizes}
\alias{model_output_sizes}
\title{Retrieve model's output dimensions.}
\usage{
model_output_sizes(url, name, config = jsonlite::fromJSON("{}"))
}
\arguments{
\item{url}{URL the model is running at.}

\item{name}{Name of the desired model}

\item{config}{Model-specific configuration options.}
}
\value{
List of output dimensions.
}
\description{
Retrieve model's output dimensions.
}
